rename util directory for webapp, minor cleanup
